+++ src/coral_fans/functions/prof/Prof.h
@@ -109,4 +109,5 @@ struct Profiler {
     void        stop();
 };
 
+void hookTick(bool);
 } // namespace coral_fans::functions