+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.1] - 2025-4-22
+
+### Added
+
++ Add command FreeCamera (Thanks for GroupMountain)
++ Add command func/self PortalDisabled
+
+### Fixed
+
++ Fix a bug about the permission level of commmand minerule
+
 ## [2.2.0] - 2025-3-7
 
 ### Added
@@ -18,6 +29,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 + Remove command tick query
+
+## [2.1.1] - 2025-4-20
+
+### Added
+
++ Add command FreeCamera (Thanks for GroupMountain)
++ Add command func/self PortalDisabled
+
+### Fixed
+
++ Fix a bug that the village function may crash BDS when village is removed
++ Fix a bug about the permission level of commmand minerule
 
 ## [2.1.0] - 2025-2-25